Revert to origin cmd_ctl_start and use new cmd_ctl_start_osx
We can combine later but this is potentially too dangerous to change at
this moment?
